- refactored file parser implementation
