+++ app/Services/Import/FileParsers/Csv.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Services\Import\FileParsers;
+
+use App\Services\Import\Contracts\FileParserInterface;
+
+class Csv implements FileParserInterface
+{
+    //
+}
