+++ app/Services/Import/FileParsers/Xml.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Services\Import\FileParsers;
+
+use App\Services\Import\Contracts\FileParserInterface;
+
+class Excel implements FileParserInterface
+{
+    //
+}
