--- app/Services/Import/FileReaders/Csv.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace App\Services\Import\FileReaders;
-
-use App\Services\Import\Contracts\FileReaderInterface;
-
-class Csv implements FileReaderInterface
-{
-    //
-}
